# packages/api-types/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AgriMinds AI-DREWS API",
    "description": "AI-Enabled Drought Early Warning and Climate-Resilient Decision Support (AI-DREWS, Choke Mountain Watershed).\n\nSign in at `POST /api/v1/auth/login` and send the access token as `Authorization: Bearer <token>`. Every forecast response carries a `provenance` block describing the model version, training data source and issue month.",
    "version": "1.1.0"
  },
  "paths": {
...
    "/api/v1/drought/map": {
      "get": {
        "tags": [
          "Drought Early Warning"
        ],
        "summary": "Gridded drought probability for one lead",
        "operationId": "drought_map_api_v1_drought_map_get",
        "security": [
          {
            "APIKeyHeader": []
          },
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "lead_month",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 3,
              "minimum": 1,
              "description": "Forecast lead time (1, 2 or 3 months)",
              "default": 1,
              "title": "Lead Month"
            },
            "description": "Forecast lead time (1, 2 or 3 months)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DroughtMapResponse"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "503": {
            "description": "Model unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/drought/watershed": {
      "get": {
        "tags": [
          "Drought Early Warning"
        ],
        "summary": "The surveyed catchment outline and which grid cells fall inside it",
        "description": "Lets a map draw the real catchment instead of a bare rectangle, and grey out the cells that are not part of it. The outline is simplified for display; the server keeps the full-precision version for deciding whether a plot is inside.",
        "operationId": "watershed_api_v1_drought_watershed_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WatershedBoundary"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Model unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          },
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/drought/cell": {
      "get": {
        "tags": [
          "Drought Early Warning"
        ],
        "summary": "Risk for one cell by row/col or GPS",
        "operationId": "cell_risk_api_v1_drought_cell_get",
        "security": [
          {
            "APIKeyHeader": []
          },
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "lead_month",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 3,
              "minimum": 1,
              "description": "Forecast lead time (1, 2 or 3 months)",
              "default": 1,
              "title": "Lead Month"
            },
            "description": "Forecast lead time (1, 2 or 3 months)"
          },
          {
            "name": "row",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "title": "Row"
            }
          },
          {
            "name": "col",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "title": "Col"
            }
          },
          {
            "name": "latitude",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number",
                  "maximum": 90,
                  "minimum": -90
                },
                {
                  "type": "null"
                }
              ],
              "title": "Latitude"
            }
          },
          {
            "name": "longitude",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number",
                  "maximum": 180,
                  "minimum": -180
                },
                {
                  "type": "null"
                }
              ],
              "title": "Longitude"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CellRiskResponse"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "503": {
            "description": "Model unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Drought Early Warning"
        ],
        "summary": "Risk for one cell (JSON body)",
        "operationId": "cell_risk_post_api_v1_drought_cell_post",
        "security": [
          {
            "APIKeyHeader": []
          },
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CellRiskQuery"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CellRiskResponse"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "503": {
            "description": "Model unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/analytics/powerbi/status": {
      "get": {
        "tags": [
          "Analytics"
        ],
        "summary": "Whether Power BI embedding is configured on this deployment",
        "description": "Lets the client show the analytics tab only when it will work, instead of rendering a broken frame. Never returns any credential.",
        "operationId": "powerbi_status_api_v1_analytics_powerbi_status_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PowerBiStatus"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Role not allowed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/analytics/powerbi/embed-token": {
      "get": {
        "tags": [
          "Analytics"
        ],
        "summary": "Short-lived token for rendering the report in the browser",
        "description": "Issued per viewer. Where the dataset defines row-level security, a development agent's token is scoped to their own woreda by Power BI, not by the client.",
        "operationId": "powerbi_embed_token_api_v1_analytics_powerbi_embed_token_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmbedConfig"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Role not allowed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "Power BI did not respond",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Embedding is not configured",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/analytics/connection": {
      "get": {
        "tags": [
          "Analytics"
        ],
        "summary": "Direct SQL connection details for building your own reports (administrators)",
        "description": "Returns where the read-only `analytics` schema lives and which views it offers. Credentials are issued separately: this endpoint never returns a password.",
        "operationId": "analytics_connection_api_v1_analytics_connection_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalyticsConnection"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Role not allowed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/analytics/connection.pbids": {
      "get": {
        "tags": [
          "Analytics"
        ],
        "summary": "Power BI Desktop connection file (administrators)",
        "description": "Opens Power BI Desktop straight onto the analytics schema. Contains no credentials.",
        "operationId": "analytics_pbids_api_v1_analytics_connection_pbids_get",
        "responses": {
          "200": {
            "description": "A .pbids file",
            "content": {
              "application/json": {}
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Role not allowed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CellRiskQuery": {
        "properties": {
          "lead_month": {
            "type": "integer",
            "maximum": 3.0,
            "minimum": 1.0,
            "title": "Lead Month",
            "default": 1
          },
          "row": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Row"
          },
          "col": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Col"
          },
          "latitude": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 90.0,
                "minimum": -90.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Latitude"
          },
          "longitude": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 180.0,
                "minimum": -180.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Longitude"
          }
        },
        "type": "object",
        "title": "CellRiskQuery",
        "description": "Either (row, col) or (latitude, longitude). Neither -> watershed centre cell."
      },
      "CellRiskResponse": {
        "properties": {
          "row": {
            "type": "integer",
            "title": "Row"
          },
          "col": {
            "type": "integer",
            "title": "Col"
          },
          "latitude": {
            "type": "number",
            "title": "Latitude"
          },
          "longitude": {
            "type": "number",
            "title": "Longitude"
          },
          "probability": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Probability"
          },
          "risk_level": {
            "type": "string",
            "enum": [
              "Low",
              "Moderate",
              "High",
              "Severe"
            ],
            "title": "Risk Level"
          },
          "pdsi": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pdsi",
            "description": "Observed self-calibrated Palmer index for the issue month"
          },
          "pdsi_category": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "Extremely wet",
                  "Very wet",
                  "Moderately wet",
                  "Normal",
                  "Moderately dry",
                  "Very dry",
                  "Extremely dry"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Pdsi Category",
            "description": "Drought intensity band for `pdsi` (Table 2)"
          },
          "in_watershed": {
            "type": "boolean",
            "title": "In Watershed",
            "description": "Whether this cell's centre falls inside the surveyed catchment. The grid is a rectangle over a catchment that is not one, so some cells are outside it.",
            "default": true
          },
          "lead_month": {
            "type": "integer",
            "title": "Lead Month"
          },
          "provenance": {
            "$ref": "#/components/schemas/ForecastProvenance"
          }
        },
        "type": "object",
        "required": [
          "row",
          "col",
          "latitude",
          "longitude",
          "probability",
          "risk_level",
          "lead_month",
          "provenance"
        ],
        "title": "CellRiskResponse"
      },
...
      "DroughtMapResponse": {
        "properties": {
          "lead_month": {
            "type": "integer",
            "title": "Lead Month"
          },
          "target_date": {
            "type": "string",
            "title": "Target Date",
            "description": "Target month, e.g. 'July 2026'"
          },
          "issued_date": {
            "type": "string",
            "title": "Issued Date",
            "description": "Issue month, e.g. 'June 2026'"
          },
          "grid_shape": {
            "prefixItems": [
              {
                "type": "integer"
              },
              {
                "type": "integer"
              }
            ],
            "type": "array",
            "maxItems": 2,
            "minItems": 2,
            "title": "Grid Shape"
          },
          "bbox": {
            "prefixItems": [
              {
                "type": "number"
              },
              {
                "type": "number"
              },
              {
                "type": "number"
              },
              {
                "type": "number"
              }
            ],
            "type": "array",
            "maxItems": 4,
            "minItems": 4,
            "title": "Bbox",
            "description": "lon_min, lat_min, lon_max, lat_max"
          },
          "mean_probability": {
            "type": "number",
            "title": "Mean Probability"
          },
          "min_probability": {
            "type": "number",
            "title": "Min Probability"
          },
          "max_probability": {
            "type": "number",
            "title": "Max Probability"
          },
          "probabilities": {
            "items": {
              "items": {
                "type": "number"
              },
              "type": "array"
            },
            "type": "array",
            "title": "Probabilities"
          },
          "cells": {
            "items": {
              "$ref": "#/components/schemas/GridCellRisk"
            },
            "type": "array",
            "title": "Cells"
          },
          "conditions": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ObservedConditions"
              },
              {
                "type": "null"
              }
            ],
            "description": "Observed Sc-PDSI; null when the record carries no mean temperature"
          },
          "provenance": {
            "$ref": "#/components/schemas/ForecastProvenance"
          }
        },
        "type": "object",
        "required": [
          "lead_month",
          "target_date",
          "issued_date",
          "grid_shape",
          "bbox",
          "mean_probability",
          "min_probability",
          "max_probability",
          "probabilities",
          "cells",
          "provenance"
        ],
        "title": "DroughtMapResponse"
      },
      "EmbedConfig": {
        "properties": {
          "report_id": {
            "type": "string",
            "title": "Report Id"
          },
          "embed_url": {
            "type": "string",
            "title": "Embed Url"
          },
          "access_token": {
            "type": "string",
            "title": "Access Token",
            "description": "Short-lived embed token, not an AAD token"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At"
          },
          "scope": {
            "type": "string",
            "title": "Scope",
            "description": "Which rows this token may read, in plain words"
          },
          "rls_applied": {
            "type": "boolean",
            "title": "Rls Applied"
          }
        },
        "type": "object",
        "required": [
          "report_id",
          "embed_url",
          "access_token",
          "expires_at",
          "scope",
          "rls_applied"
        ],
        "title": "EmbedConfig",
        "description": "Everything the browser needs to render a Power BI report, and nothing more.\n\nThe client secret stays on the server; what goes out is a short-lived token scoped to one\nreport and, where row-level security is configured, to one viewer's district."
      },
...
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PowerBiStatus": {
        "properties": {
          "configured": {
            "type": "boolean",
            "title": "Configured"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason",
            "description": "Why embedding is unavailable, when it is"
          },
          "workspace_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Workspace Id"
          },
          "report_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Report Id"
          },
          "rls_role": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rls Role"
          }
        },
        "type": "object",
        "required": [
          "configured"
        ],
        "title": "PowerBiStatus"
      },
...
      "WatershedBoundary": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "source": {
            "type": "string",
            "title": "Source",
            "description": "File the outline was surveyed from"
          },
          "area_km2": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Area Km2"
          },
          "bbox": {
            "prefixItems": [
              {
                "type": "number"
              },
              {
                "type": "number"
              },
              {
                "type": "number"
              },
              {
                "type": "number"
              }
            ],
            "type": "array",
            "maxItems": 4,
            "minItems": 4,
            "title": "Bbox",
            "description": "lon_min, lat_min, lon_max, lat_max"
          },
          "geometry": {
            "additionalProperties": true,
            "type": "object",
            "title": "Geometry",
            "description": "GeoJSON Polygon, simplified for display"
          },
          "grid": {
            "$ref": "#/components/schemas/WatershedGrid"
          }
        },
        "type": "object",
        "required": [
          "name",
          "source",
          "area_km2",
          "bbox",
          "geometry",
          "grid"
        ],
        "title": "WatershedBoundary",
        "description": "The surveyed catchment outline, for drawing the real shape instead of a rectangle."
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "description": "JWT access token from POST /auth/login",
        "scheme": "bearer"
      },
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  }
}